PRSD-NONE: Add abstract base page and enrich function
