+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/AlreadyRegisteredPage.kt
@@ -1,9 +1,7 @@
 package uk.gov.communities.prsdb.webapp.forms.pages
 
-import org.springframework.validation.Validator
 import org.springframework.web.servlet.ModelAndView
 import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
-import uk.gov.communities.prsdb.webapp.forms.journeys.PageData
 import uk.gov.communities.prsdb.webapp.helpers.JourneyDataHelper
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.FormModel
 import kotlin.reflect.KClass
@@ -13,18 +11,14 @@ class AlreadyRegisteredPage(
     templateName: String,
     content: Map<String, Any>,
     private val selectedAddressPathSegment: String,
-) : Page(formModel, templateName, content) {
-    override fun getModelAndView(
-        validator: Validator,
-        pageData: PageData?,
-        prevStepUrl: String?,
+) : AbstractPage(formModel, templateName, content) {
+    override fun enrichModel(
+        modelAndView: ModelAndView,
         journeyData: JourneyData?,
-    ): ModelAndView {
-        val modelAndView = super.getModelAndView(validator, pageData, prevStepUrl)
+    ) {
         modelAndView.addObject(
             "singleLineAddress",
             JourneyDataHelper.getFieldStringValue(journeyData!!, selectedAddressPathSegment, "address"),
         )
-        return modelAndView
     }
 }
