+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/LaUserRegistrationCheckAnswersPage.kt
@@ -1,10 +1,8 @@
 package uk.gov.communities.prsdb.webapp.forms.pages
 
-import org.springframework.validation.Validator
 import org.springframework.web.servlet.ModelAndView
 import uk.gov.communities.prsdb.webapp.exceptions.PrsdbWebException
 import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
-import uk.gov.communities.prsdb.webapp.forms.journeys.PageData
 import uk.gov.communities.prsdb.webapp.forms.steps.RegisterLaUserStepId
 import uk.gov.communities.prsdb.webapp.helpers.LaUserRegistrationJourneyDataHelper
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.FormModel
@@ -17,13 +15,11 @@ class LaUserRegistrationCheckAnswersPage(
     templateName: String,
     content: Map<String, Any>,
     private val invitationService: LocalAuthorityInvitationService,
-) : Page(formModel, templateName, content) {
-    override fun getModelAndView(
-        validator: Validator,
-        pageData: PageData?,
-        prevStepUrl: String?,
+) : AbstractPage(formModel, templateName, content) {
+    override fun enrichModel(
+        modelAndView: ModelAndView,
         journeyData: JourneyData?,
-    ): ModelAndView {
+    ) {
         journeyData!!
         val formData = mutableListOf<SummaryListRowViewModel>()
         val sessionToken = invitationService.getTokenFromSession()
@@ -55,9 +51,6 @@ class LaUserRegistrationCheckAnswersPage(
             ),
         )
 
-        val modelAndView = super.getModelAndView(validator, pageData, prevStepUrl)
         modelAndView.addObject("formData", formData)
-
-        return modelAndView
     }
 }
