+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/SelectLocalAuthorityPage.kt
@@ -1,8 +1,7 @@
 package uk.gov.communities.prsdb.webapp.forms.pages
 
-import org.springframework.validation.Validator
 import org.springframework.web.servlet.ModelAndView
-import uk.gov.communities.prsdb.webapp.forms.journeys.PageData
+import uk.gov.communities.prsdb.webapp.forms.journeys.JourneyData
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.SelectLocalAuthorityFormModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.formModels.SelectViewModel
 import uk.gov.communities.prsdb.webapp.services.LocalAuthorityService
@@ -11,17 +10,16 @@ class SelectLocalAuthorityPage(
     content: Map<String, Any>,
     displaySectionHeader: Boolean = false,
     private val localAuthorityService: LocalAuthorityService,
-) : Page(
+) : AbstractPage(
         formModel = SelectLocalAuthorityFormModel::class,
         templateName = "forms/selectLocalAuthorityForm",
         content = content,
         shouldDisplaySectionHeader = displaySectionHeader,
     ) {
-    override fun getModelAndView(
-        validator: Validator,
-        pageData: PageData?,
-        prevStepUrl: String?,
-    ): ModelAndView {
+    override fun enrichModel(
+        modelAndView: ModelAndView,
+        journeyData: JourneyData?,
+    ) {
         val localAuthoritiesSelectOptions =
             localAuthorityService.retrieveAllLocalAuthorities().map {
                 SelectViewModel(
@@ -30,9 +28,6 @@ class SelectLocalAuthorityPage(
                 )
             }
 
-        val modelAndView = super.getModelAndView(validator, pageData, prevStepUrl)
         modelAndView.addObject("selectOptions", localAuthoritiesSelectOptions)
-
-        return modelAndView
     }
 }
